mds-1435 permit document pre-work clean up (#541)
* merging mine_exp_doc and exp_doc models.

* full change of merging models.

* wrapping due date method.

* clean up due date cal method.

* documents created all or none.
